WIP: Update "weekday next" logic
The format "weekday next" (`replan sat`) is very confusing, since it does remove the replan; this should be a separate functionality (e.g. `n`oreplan, which would play nicely when using `sn` to schedule a onceoff event in the future; `o`nceoff could also be another good name, which may also imply `s`kip).

The problem is what to do with weekdays with extra specifiers, like `sat+`; probably, the `next` field could accept only some formats.

It's not very clear the logic of the conditional. This case:

    replan_data.interval.nil? && (replan_data.update_full.nil?)

is not covered.

Maybe it was intended to be:

    (replan_data.interval.nil? && replan_data.next.nil?) || replan_data.update_full.nil?

As first thing, it should be understood what the `update_full.nil?` condition does.